polling can run more often

+++ .env/thin/lib/status.pl
@@ -38,17 +38,23 @@ if (@ARGV) {
         }
     }
     elsif ( $command eq "poll" ) {
-        my $act = "start";
-        if ( system("ping -c1 -w5 shelf > /dev/null 2>&1") == 0 ) {
-            unlink $no_net if -e $no_net;
-        }
-        else {
-            $act = "stop";
-            system("touch $no_net");
-        }
-        system( 'systemctl --user ' . $act . ' drudge-session@messaging' );
-        exit 0 if -e $no_net;
         chomp( my $cache = `drudge mktemp polling` ) or die "no tempdir";
+        my $check_name = "check";
+        my $checking   = "$cache/$check_name";
+        system("find $checking -type f -name '$check_name' -mmin +1 -delete");
+        if ( !-e $checking ) {
+            my $act = "start";
+            if ( system("ping -c1 -w5 shelf > /dev/null 2>&1") == 0 ) {
+                unlink $no_net if -e $no_net;
+            }
+            else {
+                $act = "stop";
+                system("touch $no_net");
+            }
+            system( 'systemctl --user ' . $act . ' drudge-session@messaging' );
+            system("touch $checking");
+        }
+        exit 0 if -e $no_net;
         $cache = "$cache/notify";
         system("drudge messaging.reader > $cache");
         if ( -s $cache ) {
